app/requirements.txt
# Core Framework
streamlit>=1.28.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlite3  # Built-in with Python

# OCR and Image Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Date and Time
python-dateutil>=2.8.2

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
regex>=2023.8.8

# Optional: Type checking
# mypy>=1.5.0

# Optional: Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
